+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Possible log types:
 ### Unreleased
 
 - [added] Allow specifying `RenderingType` for file messages
+- [added] Allow specifying media metadata for file messages
 - [changed] The API for `E2eApi::encrypt_file_msg` has changed
 - [changed] You now need to import `std::str::FromStr` to directly access
   `BlobId::from_str` or `RecipientKey::from_str`